AddCorpus successfully divided into Presentational and container
