+++ src/Elements/dummypapers.js
@@ -1,0 +1,487 @@
+const dummypapers = `%0 Journal Article
+%A Crawley, D.
+%A Chandra, A.
+%A Loda, M.
+%A Gillett, C.
+%A Cathcart, P.
+%A Challacombe, B.
+%A Cook, G.
+%A Cahill, D.
+%A Cahill, F.
+%A Santa Olalla, A.
+%A George, G.
+%A Rudman, S.
+%A Van Hemelrijck, M.
+%D 2016
+%T Metformin and longevity (METAL): A window of opportunity study investigating the biological effects of metformin in localised prostate cancer
+%B Annals of Oncology. Conference: 41st European Society for Medical Oncology Congress, ESMO 2016. Denmark. Conference Start: 20161007. Conference End: 20161011
+%V 27
+%N no pagination
+%! Metformin and longevity (METAL): A window of opportunity study investigating the biological effects of metformin in localised prostate cancer
+%R http://dx.doi.org/10.1093/annonc/mdw372.49
+%M CN-01296247
+%X Background: Metformin (1,1-dimethylbiguanide hydrochloride) is a biguanide oral hypoglycaemic agent commonly used for the treatment of type 2 diabetes mellitus. In addition to it's anti-diabetic effect, metformin has also been associated with a reduced risk of cancer incidence of a number of solid tumour including prostate cancer (PCa). However, the underlying biological mechanisms for these observations have not been fully characterised in PCa. One hypothesis is that the indirect insulin lowering effect may have an anti-neoplastic effect as elevated insulin and insulin like growth factor -1 (IGF-1) levels play a role in PCa development and progression. In addition, metformin is a potent activator of activated protein kinase (AMPK) which in turn inhibits the mammalian target of rapamycin (mTOR) and other signal transduction mechanisms . These direct effects can lead to reduced cell proliferation. Given its wide availability and tolerable side effect profile, metformin represents an attractive potential therapeutic option for men with PCa. Hence, the need for a clinical trial investigating its biological mechanisms in PCa. Trial design: METAL is a multi-centre, randomized, placebo-controlled, double-blind, window of opportunity study investigating the biological mechanism of metformin in prostate cancer. 180 patients with newly-diagnosed, localised PCa scheduled for radical prostatectomy will be randomised 1:1 to receive metformin (1g bd) or placebo for four weeks (+/- 1 week) prior to prostatectomy. Tissue will be collected from both diagnostic biopsy and prostatectomy specimens. The primary endpoint is the difference in expression levels of markers of the FASN/AMPK pathway pre and post treatment between the placebo and metformin arms. Secondary endpoints include the difference in expression levels of indicators of proliferation (ki67 and TUNEL) pre and post treatment between the placebo and metformin arms. METAL is currently open to recruitment at Guy's and St Thomas' Hospital and the RoyalMarsden Hospital, London.
+%~ CENTRAL
+%G eng
+
+
+
+%0 Journal Article
+%A Czech, M.
+%A Rdzanek, E.
+%A Paweska, J.
+%A Adamowicz-Sidor, O.
+%A Niewada, M.
+%A Jakubczyk, M.
+%D 2015
+%T Drug-related risk of severe hypoglycaemia in observational studies: A systematic review and meta-analysis
+%B BMC Endocrine Disorders
+%V 15 (1) (no pagination)
+%N 57
+%8 October 12
+%! Drug-related risk of severe hypoglycaemia in observational studies: A systematic review and meta-analysis
+%@ 1472-6823
+%R http://dx.doi.org/10.1186/s12902-015-0052-z
+%M 606362438
+%K Antidiabetic medications
+Diabetes mellitus
+Insulin regimens
+Observational studies
+Severe hypoglycaemia
+article
+human
+hypoglycemia/ep [Epidemiology]
+hypoglycemia/si [Side Effect]
+incidence
+insulin dependent diabetes mellitus/dt [Drug Therapy]
+insulin pump
+non insulin dependent diabetes mellitus/dt [Drug Therapy]
+observational study
+risk assessment
+risk factor
+systematic review
+biphasic insulin/ae [Adverse Drug Reaction]
+biphasic insulin/dt [Drug Therapy]
+dipeptidyl peptidase IV inhibitor/dt [Drug Therapy]
+glitazone derivative/dt [Drug Therapy]
+glucagon like peptide 1/dt [Drug Therapy]
+human insulin/ae [Adverse Drug Reaction]
+human insulin/dt [Drug Therapy]
+insulin derivative/ae [Adverse Drug Reaction]
+insulin derivative/dt [Drug Therapy]
+long acting insulin/ae [Adverse Drug Reaction]
+long acting insulin/dt [Drug Therapy]
+metformin/dt [Drug Therapy]
+oral antidiabetic agent/cb [Drug Combination]
+oral antidiabetic agent/dt [Drug Therapy]
+sulfonylurea derivative/ae [Adverse Drug Reaction]
+sulfonylurea derivative/cb [Drug Combination]
+sulfonylurea derivative/dt [Drug Therapy]
+%X Background: Diabetes mellitus (DM) leads to multiple complications, including severe hypoglycaemia events (SHEs). SHEs can impact a patient's quality of life and compliance and may directly result in additional costs to the health care system. The aim of this review was to evaluate the risk of severe hypoglycaemia in patients with type 1 (T1) and 2 (T2) DM as observed in everyday clinical practice for various drug regimens. Methods: We conducted a systematic review of observational (retrospective or prospective) studies in the MEDLINE, Embase, and Cochrane Library databases that covered at least 100 children or adults with T1/T2 DM. In T1 DM, basal-bolus/pre-mix insulin (human or analogue) and insulin pump were reviewed, and in T2 DM, basal-bolus/pre-mix insulin (human or analogue), oral antidiabetic drugs supported with basal insulin (human or analogue), sulfonylureas in monotherapy, and combined oral treatment were reviewed. In order to estimate SHE rates, we extracted data on the time horizon of the study, number of patients, number of SHEs, and number of patients experiencing at least one SHE. We used a random effects model to estimate the annual SHE rate. We considered the risk for other antidiabetic medications in T2 DM to be negligible and the results of our main review yielded no observational data for premixes in T1 DM so they were assessed based on relative rates taken from additional systematic reviews. The study, being a desk research, did not involve any human subjects (including human material or human data) and no ethical committee approval was asked for. For the same reason there was no need to collect informed consent for participation in the study. Results: We identified 76 observational studies encompassing 707,722.30 patient-years. The estimated annual SHE rate varied from 0.168 (95 % CI 0.123-0.237) for insulin pump up to 1.628 for biphasic human insulin in T1 DM patients, and from 0.0035 for oral antidiabetic drugs up to 0.554 (95 % CI 0.157-7.534) for basal-bolus with human insulin in T2 DM patients. Conclusions: Our review indicates that SHE rates differ between patients depending on treatment regimen. However, SHEs are also driven by other factors. Proper modelling techniques are needed to use various types of information in published studies. Copyright © 2015 Czech et al.
+%U http://www.biomedcentral.com/bmcendocrdisord/
+http://ovidsp.ovid.com/ovidweb.cgi?T=JS&CSC=Y&NEWS=N&PAGE=fulltext&D=emed17&AN=606362438
+%+ (Czech) Novo Nordisk Pharma sp. z o.o, Warsaw, Poland (Czech) Medical University of Warsaw, Department of Pharmacoeconomics, Warsaw, Poland (Czech) Warsaw University of Technology, Business School, Warsaw, Poland (Rdzanek, Paweska, Adamowicz-Sidor) HealthQuest spolka z ograniczona odpowiedzialnoscia Sp. K, Warsaw, Poland (Niewada) Medical University of Warsaw, Department of Experimental and Clinical Pharmacology, Warsaw, Poland (Jakubczyk) Warsaw School of Economics, Decision Analysis and Support Unit, Al. Niepodleglosci 162, Warsaw 02-554, Poland
+M. Jakubczyk, Warsaw School of Economics, Decision Analysis and Support Unit, Al. Niepodleglosci 162, Warsaw 02-554, Poland
+%~ Embase
+%W Ovid Technologies
+%G English
+
+
+
+%0 Journal Article
+%A Davidson, J. A.
+%A Schernthaner, G.
+%A Hieronymus, L.
+%A Jodon, H.
+%A Vijapurkar, U.
+%A Meininger, G.
+%A Canovatchel, W.
+%D 2015
+%T Canagliflozin is superior to sitagliptin in reducing both a1c and body weight in patients with type 2 diabetes mellitus
+%B Endocrine Reviews. Conference: 97th Annual Meeting and Expo of the Endocrine Society, ENDO 2015. United States. Conference Start: 20150305. Conference End: 20150308
+%V 36
+%N no pagination
+%! Canagliflozin is superior to sitagliptin in reducing both a1c and body weight in patients with type 2 diabetes mellitus
+%M CN-01293720
+%X Many patients with type 2 diabetes mellitus (T2DM) are obese or overweight, so treatments that improve glycemic control and reduce body weight may be beneficial. Canagliflozin (CANA), a sodium glucose co-transporter 2 (SGLT2) inhibitor approved for the treatment of T2DM, inhibits renal glucose reabsorption by lowering the renal threshold for glucose (RTG) and increasing urinary glucose excretion (UGE), resulting in lowered plasma glucose levels and a net caloric loss. CANA has been shown to provide reductions in A1C and body weight compared with sitagliptin (SITA) in patients with T2DM inadequately controlled on metformin (MET) or MET plus sulfonylurea (SU; CANA 300 mg only) in two 52-week, randomized, double-blind, active-controlled, Phase 3 studies. This analysis evaluated the proportion of patients who responded to treatment in terms of both A1C and body weight. In Study 1 (N = 1,284; mean baseline A1C, 7.9%; body weight, 87.2 kg), least squares (LS) mean changes from baseline in A1C at Week 52 were -0.73%, -0.88%, and -0.73% with CANA 100 and 300 mg and SITA 100 mg, respectively; LS mean percent changes in body weight were -3.8%, -4.2%, and -1.3%, respectively. In Study 2 (N = 755; mean baseline A1C, 8.1%; body weight, 88.3 kg), LS mean changes from baseline in A1C were -1.03% and -0.66% at Week 52, and LS mean percent changes in body weight were -2.5% and 0.3% with CANA 300 mg and SITA 100 mg, respectively. In Study 1, the proportion of patients with reductions in A1C were 77.8%, 84.7%, and 80.5% with CANA 100 and 300 mg and SITA 100 mg, respectively; 82.7%, 84.7%, and 58.8% of patients had reductions in body weight. A greater proportion of patients in Study 1 had reductions in both A1C and body weight with CANA 100 and 300 mg versus SITA 100 mg (67.7%, 74.7%, and 50.6%, respectively; differences [95% confidence interval (CI)] of 17.1% [9.7, 24.5] and 24.2% [17.1, 31.3], respectively). In Study 2, the proportion of patients with reductions in A1C with CANA 300 mg and SITA 100 mg were 86.6% and 74.2%, respectively; 78.6% and 47.1% of patients had reductions in body weight. Similar to Study 1, a greater proportion of patients in Study 2 had reductions in both A1C and body weight with CANA 300 mg (70.6%) versus SITA 100 mg (37.8%; difference [95% CI] of 32.8% [25.7, 39.8]). The overall incidence of adverse events (AEs) was similar across treatment groups in both studies; CANA was associated with increased rates of genital mycotic infections and osmotic diuresis-related AEs compared with SITA. In Study 1, the incidence of documented hypoglycemia (<70 mg/dL) was 7%, 7%, and 4% with CANA 100 and 300 mg and SITA 100 mg, respectively; in Study 2, the incidence of documented hypoglycemia was 43% and 41% with CANA 300 mg and SITA 100 mg, respectively. In conclusion, CANA provided greater attainment of reduction in both A1C and body weight compared with SITA at 52 weeks and was generally well tolerated in patients with T2DM as add-on to MET or MET plus SU.
+%~ CENTRAL
+%G eng
+
+
+
+%0 Journal Article
+%A Davies, M. J.
+%A Merton, K.
+%A Vijapurkar, U.
+%A Yee, J.
+%A Qiu, R.
+%D 2017
+%T Efficacy and safety of canagliflozin in patients with type 2 diabetes based on history of cardiovascular disease or cardiovascular risk factors: A post hoc analysis of pooled data
+%B Cardiovascular Diabetology
+%V 16 (1) (no pagination)
+%N 40
+%8 21 Mar
+%! Efficacy and safety of canagliflozin in patients with type 2 diabetes based on history of cardiovascular disease or cardiovascular risk factors: A post hoc analysis of pooled data
+%@ 1475-2840
+%R http://dx.doi.org/10.1186/s12933-017-0517-7
+%M 614941180
+%K Canagliflozin
+Cardiovascular disease
+Risk factors
+SGLT2 inhibitor
+Type 2 diabetes mellitus
+adult
+aged
+article
+body weight
+cardiovascular risk
+combination drug therapy
+controlled clinical trial
+controlled study
+drug dose comparison
+drug efficacy
+drug safety
+drug use
+drug withdrawal
+female
+human
+hypertension
+incidence
+major clinical study
+male
+medical history
+monotherapy
+non insulin dependent diabetes mellitus/dt [Drug Therapy]
+phase 3 clinical trial
+post hoc analysis
+risk assessment
+systolic blood pressure
+canagliflozin/ct [Clinical Trial]
+canagliflozin/cb [Drug Combination]
+canagliflozin/cm [Drug Comparison]
+canagliflozin/do [Drug Dose]
+canagliflozin/dt [Drug Therapy]
+hemoglobin A1c/ec [Endogenous Compound]
+hydroxymethylglutaryl coenzyme A reductase inhibitor
+metformin/cb [Drug Combination]
+metformin/dt [Drug Therapy]
+pioglitazone/cb [Drug Combination]
+pioglitazone/dt [Drug Therapy]
+placebo
+sulfonylurea/cb [Drug Combination]
+sulfonylurea/dt [Drug Therapy]
+%X Background: Treatment of patients with type 2 diabetes mellitus (T2DM) and a history of cardiovascular (CV) disease or CV risk factors may present clinical challenges due to the presence of comorbid conditions and the use of concomitant medications. The sodium glucose co-transporter 2 inhibitor, canagliflozin, has been shown to improve glycaemic control and reduce body weight and blood pressure (BP) with a favourable tolerability profile in a broad range of patients with T2DM. This post hoc analysis assessed the efficacy and safety of canagliflozin in patients with T2DM based on CV disease history or CV risk factors. Methods: Analyses were based on pooled data from four 26-week, placebo-controlled, Phase 3 studies that evaluated canagliflozin 100 and 300mg in patients with T2DM (N=2313; mean HbA1c, 8.0%; body weight, 89kg; systolic BP, 128mmHg). Changes from baseline to week26 in HbA1c, body weight, and systolic BP were assessed based on history of CV disease, history of hypertension, baseline statin use, and number of CV risk factors. Safety was assessed based on adverse event (AE) reports. Results: At week 26, both canagliflozin doses lowered HbA1c, body weight, and systolic BP compared with placebo in patients with and without CV disease history or risk factors. Placebo-subtracted HbA1c reductions with canagliflozin 100 and 300mg were similar in patients with a history of CV disease (0.95 and 1.07%) versus no history of CV disease (0.71 and 0.90%), history of hypertension (0.72 and 0.89%) versus no history of hypertension (0.73 and 0.95%), baseline statin use (0.77 and 0.99%) versus no statin use (0.69 and 0.85%), and 0-1 CV risk factor (0.72 and 0.87%) versus >2 CV risk factors (0.74 and 1.02%). Similar body weight and systolic BP reductions were seen with canagliflozin versus placebo across subgroups. The incidence of AEs, AEs leading to discontinuation, and serious AEs was similar across subgroups. Conclusions: The efficacy and safety of canagliflozin were generally consistent across subgroups of patients with T2DM and varying degrees of CV disease history or risk factors. Trial registration numbers and dates ClinicalTrials.gov: NCT01081834, 4 March 2010; NCT01106625, 1 April 2010; NCT01106677, 1 April 2010; NCT01106690, 1 April 2010 Copyright © 2017 The Author(s).
+%U http://www.cardiab.com/home/
+http://ovidsp.ovid.com/ovidweb.cgi?T=JS&CSC=Y&NEWS=N&PAGE=fulltext&D=emedx&AN=614941180
+%+ (Davies, Merton) Janssen Scientific Affairs, LLC, 1125 Trenton-Harbourton Road, Titusville, NJ 08560, United States (Vijapurkar, Yee, Qiu) Janssen Research and Development, LLC, 920 US Highway 202 South, Raritan, NJ 08869, United States
+M.J. Davies, Janssen Scientific Affairs LLC, 1125 Trenton-Harbourton Road, Titusville, NJ 08560, United States. E-mail: mdavies9@its.jnj.com
+%~ Embase
+%W Ovid Technologies
+%G English
+
+
+
+%0 Journal Article
+%A DeFronzo, R. A.
+%A Bergenstal, R. M.
+%A Bode, B.
+%A Kushner, R.
+%A Lewin, A. J.
+%A Skjoth, T. V.
+%A Jensen, C. B.
+%A Davies, M.
+%D 2014
+%T Effects of liraglutide 3.0 mg and 1.8 mg on body weight and cardiometabolic risk factors in overweight and obese adults with type 2 diabetes mellitus (T2DM): the scale diabetes randomized, double-blind, placebo-controlled, 56-week trial
+%B Endocrine Reviews. Conference: 96th Annual Meeting and Expo of the Endocrine Society, ENDO 2014. Chicago, IL United States. Conference Start: 20140621. Conference End: 20140624. Conference Publication: (var.pagings)
+%V 35
+%N no pagination
+%! Effects of liraglutide 3.0 mg and 1.8 mg on body weight and cardiometabolic risk factors in overweight and obese adults with type 2 diabetes mellitus (T2DM): the scale diabetes randomized, double-blind, placebo-controlled, 56-week trial
+%M CN-01267316
+%X Obesity increases risk of multiple comorbidities including T2DM and cardiovascular (CV) disease. A 5-10% weight loss can decrease risk but is difficult to achieve and maintain by lifestyle changes alone (1). Liraglutide at doses up to 1.8 mg is approved for treatment of T2DM. Four phase 3 studies for weight management (SCALE) were recently completed. This study evaluated the effects of liraglutide (3.0 and 1.8 mg) as adjunct to diet and exercise on weight loss (primary endpoint) and cardiometabolic risk factors (waist circumference, glycemia, blood pressure [BP], pulse, fasting lipids, and CV biomarkers) in overweight/obese adults with T2DM. 846 individuals (male 50%, mean age 54.9 y, BMI 37.1 kg/m2 [27.0-67.6], HbA1c 7.9%, diabetes duration 7.3 y [0.2-36.5], 11.5% on diet and exercise, 57.3% on metformin, and 31.2% on combination OADs) were randomized (2:1:1) to liraglutide 3.0 mg, 1.8 mg, or placebo for 56 weeks; all arms included a 500 kcal/day deficit diet and exercise. Clinicaltrials.gov ID: NCT01272232. At week 56, mean (least square) weight loss was 5.9%, 4.6% and 2.0% for 3.0 mg, 1.8 mg, and placebo, respectively; 3.0 mg and 1.8 mg were superior to placebo (p<0.0001); 3.0 mg was superior to 1.8 mg (p=0.0024). Weight loss was accompanied by reductions in waist circumference of 6.0, 4.9, and 2.8 cm respectively; 3.0 mg and 1.8 mg were superior to placebo (p<0.0004); 3.0 mg was superior to 1.8 mg (p=0.0224). HbA1c decreased by 1.32%, 1.13% and 0.38% from baseline with 3.0 mg, 1.8 mg, and placebo respectively (p<0.0001 vs. placebo; p=0.0125 for 3.0 mg vs. 1.8 mg). At baseline, systolic (S) /diastolic (D) BP and lipids were well controlled. SBP reductions were 3.0, 3.1, and 0.4 mmHg from baseline (~130 mmHg) for the 3 arms respectively; greater with 3.0 mg and 1.8 mg (p<0.05 vs. placebo), with no significant differences between treatments in DBP at week 56. Similar pulse changes of 2.0 and 2.2 beats/min occurred with liraglutide 3.0 and 1.8 mg vs. -1.5 beats/min for placebo (p<0.0001 vs. placebo). Liraglutide 3.0 mg reduced total cholesterol, VLDL and triglycerides (-4%, -13%, -14%, p<0.05) and increased HDL (+3%, p<0.05) vs. placebo whereas 1.8 mg had no effects. Liraglutide had no significant effects on LDL and FFA. Reductions in highsensitivity C-reactive protein occurred with liraglutide 3.0 and 1.8 mg (-27% and -25% vs. placebo, p< 0.0002). Decreases in plasminogen activator inhibitor-1 (-24%, p=0.0004) and urinary albumin/creatinine ratio (-20%, p=0.0086) occurred only with liraglutide 3.0 mg vs. placebo; No treatment effects on adiponectin occurred. Fibrinogen increased by 5% vs. placebo with 3.0 mg only (p=0.046). In conclusion, liraglutide 3.0 mg as an adjunct to diet and exercise caused significant weight loss that was associated with additional benefits for multiple cardiometabolic risk factors compared to liraglutide 1.8 mg and placebo in overweight/obese adults with T2DM.
+%~ CENTRAL
+%G eng
+
+
+
+%0 Journal Article
+%A Denig, P.
+%A Schuling, J.
+%A Haaijer-Ruskamp, F.
+%A Voorham, J.
+%D 2014
+%T Effects of a patient oriented decision aid for prioritising treatment goals in diabetes: Pragmatic randomised controlled trial
+%B BMJ (Online)
+%V 349 (no pagination)
+%N g5651
+%8 25 Sep
+%! Effects of a patient oriented decision aid for prioritising treatment goals in diabetes: Pragmatic randomised controlled trial
+%@ 0959-8146
+1756-1833
+%R http://dx.doi.org/10.1136/bmj.g5651
+%M 600055285
+%K adult
+aged
+albuminuria
+article
+automation
+blood pressure regulation
+cholesterol blood level
+controlled study
+data extraction
+diabetes mellitus/dt [Drug Therapy]
+electronic medical record
+empowerment
+female
+glycemic control
+goal attainment
+human
+intervention study
+major clinical study
+male
+middle aged
+motivation
+outcome assessment
+patient attitude
+patient decision making
+patient oriented decision aid
+prescription
+randomized controlled trial
+scoring system
+treatment planning
+albumin/ec [Endogenous Compound]
+angiotensin receptor antagonist/ct [Clinical Trial]
+antidiabetic agent/ct [Clinical Trial]
+antidiabetic agent/dt [Drug Therapy]
+antilipemic agent/ct [Clinical Trial]
+beta adrenergic receptor blocking agent/ct [Clinical Trial]
+calcium antagonist/ct [Clinical Trial]
+cholesterol/ec [Endogenous Compound]
+dipeptidyl carboxypeptidase inhibitor/ct [Clinical Trial]
+diuretic agent/ct [Clinical Trial]
+glucose/ec [Endogenous Compound]
+hemoglobin A1c/ec [Endogenous Compound]
+hydroxymethylglutaryl coenzyme A reductase inhibitor/ct [Clinical Trial]
+insulin/ct [Clinical Trial]
+insulin/dt [Drug Therapy]
+metformin/ct [Clinical Trial]
+metformin/dt [Drug Therapy]
+sulfonylurea derivative/ct [Clinical Trial]
+sulfonylurea derivative/dt [Drug Therapy]
+%X Objective: To assess the effects of a patient oriented decision aid for prioritising treatment goals in diabetes compared with usual care on patient empowerment and treatment decisions. Design: Pragmatic randomised controlled trial. Setting: 18 general practices in the north of the Netherlands. Participants: 344 patients with type 2 diabetes aged <65 years at the time of diagnosis and managed in primary care between April 2011 and August 2012: 225 were allocated to the intervention group and 119 to the usual care group. Intervention: The intervention comprised a decision aid for people with diabetes, with individually tailored risk information and treatment options for multiple risk factors. The aid was intended to empower patients to prioritise between clinical domains and to support treatment decisions. It was offered to participants before a regular diabetes check-up and to their healthcare provider during the consultation. Four different formats of the decision aid were included for additional explorative analyses. Main outcome measures: The primary outcome was the effects on patient empowerment for setting and achieving goals. The secondary outcomes were changes in the prescribing of drugs to regulate glucose, blood pressure, lipids, and albuminuria. Data were collected through structured questionnaires and automated data extraction from electronic health records during six months before and after the intervention. Results: Of all intervention participants, 103 (46%) reported to have received the basic elements of the intervention. For the primary outcome analysis, 199 intervention and 107 control patients with sufficient baseline and follow-up data could be included. The mean empowerment score increased 0.1 on a 5 point scale in the overall intervention group, which was not significantly different from that of the control group (mean difference after adjusting for baseline 0.039, 95% confidence interval -0.056 to 0.134). Lipid regulating drug treatment was intensified in 25% of intervention and 12% of control participants with increased cholesterol levels, which did not reach significance when the intervention was compared with the usual care group (odds ratio 2.54, 95% confidence interval 0.89 to 7.23). Prespecified explorative analyses showed that this effect was significant for the printed version of the decision aid in comparison to usual care (3.90, 1.29 to 11.80). No relevant or significant changes were seen for other treatments. Conclusion: We found no evidence that the patient oriented treatment decision aid improves patient empowerment by an important amount. The aid was not used to its full extent in a substantial number of participants. Trial registration: Dutch trial register NTR1942. Copyright © 2014, BMJ Publishing Group. All rights reserved.
+%U http://www.bmj.com/content/bmj/349/bmj.g5651.full.pdf
+http://ovidsp.ovid.com/ovidweb.cgi?T=JS&CSC=Y&NEWS=N&PAGE=fulltext&D=emed16&AN=600055285
+%+ (Denig, Haaijer-Ruskamp, Voorham) Clinical Pharmacy and Pharmacology, University Medical Center Groningen, PO Box 30.001, Groningen, Netherlands (Denig, Schuling, Haaijer-Ruskamp, Voorham) University of Groningen, University Medical Center Groningen, Groningen, Netherlands
+P. Denig, Clinical Pharmacy and Pharmacology, University Medical Center Groningen, PO Box 30.001, Groningen, Netherlands
+%~ Embase
+%W Ovid Technologies
+%G English
+
+
+
+%0 Journal Article
+%A Derosa, G.
+%A Mugellini, A.
+%A Pesce, R. M.
+%A D'Angelo, A.
+%A Maffioli, P.
+%D 2015
+%T A study about the relevance of adding acetylsalicylic acid in primary prevention in subjects with type 2 diabetes mellitus: Effects on some new emerging biomarkers of cardiovascular risk
+%B Cardiovascular Diabetology
+%V 14 (1) (no pagination)
+%N 95
+%! A study about the relevance of adding acetylsalicylic acid in primary prevention in subjects with type 2 diabetes mellitus: Effects on some new emerging biomarkers of cardiovascular risk
+%@ 1475-2840
+%R http://dx.doi.org/10.1186/s12933-015-0254-8
+%M 612320487
+%K Acetylsalicylic acid
+Amlodipine
+Cardiovascular risk
+Hypertension
+New emerging biomarkers
+Primary prevention
+adult
+anamnesis
+article
+blood pressure measurement
+controlled study
+diastolic blood pressure
+diet therapy
+disease severity
+double blind procedure
+drug dose titration
+epigastric pain/si [Side Effect]
+epistaxis/si [Side Effect]
+exercise
+female
+human
+hypertension/dt [Drug Therapy]
+hypertension/th [Therapy]
+Italy
+major clinical study
+male
+non insulin dependent diabetes mellitus/dt [Drug Therapy]
+non insulin dependent diabetes mellitus/pc [Prevention]
+non insulin dependent diabetes mellitus/th [Therapy]
+peripheral edema/si [Side Effect]
+physical examination
+physiotherapy
+randomized controlled trial
+systolic blood pressure
+acarbose/dt [Drug Therapy]
+acetylsalicylic acid/ae [Adverse Drug Reaction]
+acetylsalicylic acid/ct [Clinical Trial]
+acetylsalicylic acid/cb [Drug Combination]
+acetylsalicylic acid/cm [Drug Comparison]
+acetylsalicylic acid/dt [Drug Therapy]
+adiponectin/ec [Endogenous Compound]
+amlodipine/ae [Adverse Drug Reaction]
+amlodipine/ct [Clinical Trial]
+amlodipine/cb [Drug Combination]
+amlodipine/cm [Drug Comparison]
+amlodipine/dt [Drug Therapy]
+biguanide/dt [Drug Therapy]
+C reactive protein/ec [Endogenous Compound]
+CD40 ligand/ec [Endogenous Compound]
+exendin 4/dt [Drug Therapy]
+glibenclamide/dt [Drug Therapy]
+gliclazide/dt [Drug Therapy]
+glimepiride/dt [Drug Therapy]
+interleukin 1beta/ec [Endogenous Compound]
+linagliptin/dt [Drug Therapy]
+liraglutide/dt [Drug Therapy]
+metformin/dt [Drug Therapy]
+myeloperoxidase/ec [Endogenous Compound]
+pioglitazone/dt [Drug Therapy]
+repaglinide/dt [Drug Therapy]
+rosiglitazone/dt [Drug Therapy]
+saxagliptin/dt [Drug Therapy]
+sitagliptin/dt [Drug Therapy]
+sulfonylurea/dt [Drug Therapy]
+tumor necrosis factor/ec [Endogenous Compound]
+vildagliptin/dt [Drug Therapy]
+%X Aim: To evaluate the relevance of adding acetylsalicylic acid (ASA) in primary prevention in subjects with type 2 diabetes mellitus. Methods: 213 patients with type 2 diabetes mellitus and hypertension were randomized to amlodipine 5 mg, or amlodipine 5 mg + ASA 100 mg for 3 months (Phase A); then, if adequate blood pressure control was reached patients terminated the study; otherwise, amlodipine was up-titrated to 10 mg/day for further 3 months and compared to amlodipine 10 mg + ASA 100 mg (Phase B). We assessed at baseline, at the end of Phase A, and at the end of Phase B the levels of some new emerging biomarkers of cardiovascular risk including: high sensitivity C-reactive protein (Hs-CRP), adiponectin (ADN), tumor necrosis factor-alpha (TNF-alpha), interleukin-1beta (IL-1beta), myeloperoxidase (MPO), soluble CD40 ligand (sCDL40). Results: Compared to baseline, at the end of Phase A, patients treated with amlodipine 5 mg + ASA 100 mg showed a statistically significant reduction of Hs-CRP (-15.0%), TNF-alpha (-21.7%), MPO (-9.7%), and sCDL40 (-15.7%), and a statistically significant increase of ADN (+15.0%). These values were significantly better than the ones obtained with amlodipine alone. Similarly, at the end of Phase B, amlodipine 10 mg + ASA significantly lowered Hs-CRP (-18.8%), TNF-alpha (-15.0%), MPO (-9.2%), and sCDL40 (-20.0%) and increased ADN (+11.8%), with a better effect compared to amlodipine alone. Conclusion: All biomarkers considered were significantly improved by ASA addition. These data suggest that the use of ASA in primary prevention could be useful in patients with type 2 diabetes mellitus and hypertension. Copyright © 2015 Derosa et al.
+%U http://www.cardiab.com/home/
+http://ovidsp.ovid.com/ovidweb.cgi?T=JS&CSC=Y&NEWS=N&PAGE=fulltext&D=emex&AN=612320487
+%+ (Derosa, Mugellini, Pesce, D'Angelo, Maffioli) Department of Internal Medicine and Therapeutics, University of Pavia, Fondazione IRCCS Policlinico San Matteo, P.le C. Golgi, 2, Pavia 27100, Italy (Derosa) Center for the Study of Endocrine-Metabolic Pathophysiology and Clinical Research, University of Pavia, Pavia, Italy (Derosa, D'Angelo) Laboratory of Molecular Medicine, University of Pavia, Pavia, Italy (Maffioli) PhD School in Experimental Medicine, University of Pavia, Pavia, Italy
+G. Derosa, Department of Internal Medicine and Therapeutics, University of Pavia, Fondazione IRCCS Policlinico San Matteo, P.le C. Golgi, 2, Pavia 27100, Italy. E-mail: giuseppe.derosa@unipv.it
+%~ Embase
+%W Ovid Technologies
+%G English
+
+
+
+%0 Journal Article
+%A Dhindsa, S.
+%A Jialal, I.
+%D 2014
+%T Potential anti-atherosclerotic effects of dipeptidyl peptidase-4 inhibitors in type 2 diabetes mellitus topical collection on macrovascular complications in diabetes
+%B Current Diabetes Reports
+%V 14 (2) (no pagination)
+%N 463
+%8 February
+%! Potential anti-atherosclerotic effects of dipeptidyl peptidase-4 inhibitors in type 2 diabetes mellitus topical collection on macrovascular complications in diabetes
+%@ 1534-4827
+1539-0829
+%R http://dx.doi.org/10.1007/s11892-013-0463-z
+%M 52943691
+%K Antioxidant
+Atherosclerosis
+cd26
+crp
+Diabetes
+Dipeptidyl peptidase 4
+dpp-4
+Endothelial function
+Inflammation
+Lipids
+Postprandial hyperlipidemia
+t2dm
+antiatherosclerotic activity
+antiinflammatory activity
+area under the curve
+arterial wall thickness
+article
+blood vessel reactivity
+cardiovascular disease
+cardiovascular risk
+clinical trial (topic)
+drug activity
+endothelial dysfunction
+endothelial progenitor cell
+glucose blood level
+human
+mean amplitude glycemic excursion
+monocyte
+non insulin dependent diabetes mellitus
+oxidative stress
+randomized controlled trial(topic)
+risk assessment
+thrombocyte function
+triacylglycerol blood level
+3 nitrotyrosine
+chemokine receptor CCR2/ec [Endogenous Compound]
+dipeptidyl peptidase IV inhibitor/pd [Pharmacology]
+exendin 4
+glucagon like peptide 1
+hemoglobin A1c/ec [Endogenous Compound]
+I kappa B kinase beta/ec [Endogenous Compound]
+incretin/ec [Endogenous Compound]
+interleukin 10/ec [Endogenous Compound]
+interleukin 18/ec [Endogenous Compound]
+interleukin 6/ec [Endogenous Compound]
+intermediate density lipoprotein/ec [Endogenous Compound]
+metformin
+placebo
+sitagliptin
+toll like receptor 4/ec [Endogenous Compound]
+tumor necrosis factor alpha./ec [Endogenous Compound]
+very low density lipoprotein/ec [Endogenous Compound]
+vildagliptin
+%X Cardiovascular disease (CVD) is the leading cause of mortality in patients with diabetes. Pharmacotherapy that can reduce hyperglycemia and also exhibit pleiotropic effects that can result in a reduction in cardiovascular disease will be a major advance. Recently, the dipeptidyl-peptidase-4 inhibitors were introduced as ant-hyperglycemic therapy. Studies from numerous groups have reported effects that could potentially result in a reduction in CVD. Some of the drugs in this class, especially vildagliptin and sitagliptin, have been shown to reduce postprandial hyperlipidemia following a fat load, improve endothelial function as evidenced by increased forearm blood flow, and also display anti-inflammatory effects. Their effects on platelet function, blood pressure, and oxidative stress are very preliminary and need to be confirmed. Finally, they have been shown to reduce subclinical atherosclerosis by reducing carotid intimal-medial thickness. However, the final arbiter with respect to a reduction in CVD will be the ongoing clinical trials. © 2014 Springer Science+Business Media New York.
+%U http://www.springer.com
+http://ovidsp.ovid.com/ovidweb.cgi?T=JS&CSC=Y&NEWS=N&PAGE=fulltext&D=emex&AN=52943691
+%+ (Dhindsa) Division of Endocrinology and Metabolism, Texas Tech University, Permian Basin Campus, Odessa, TX 79763, United States (Jialal) Laboratory for Atherosclerosis and Metabolic Research, University of California, Davis Medical Center, 4635 Second Ave, Sacramento CA, 95817, United States (Jialal) VA Medical Center, Mather, CA 95655, United States
+I. Jialal, Laboratory for Atherosclerosis and Metabolic Research, University of California, Davis Medical Center, 4635 Second Ave, Sacramento CA, 95817, United States. E-mail: ishwarlal.jialal@ucdmc.ucdavis.edu
+%~ Embase
+%W Ovid Technologies
+%G English
+
+%0 Journal Article
+%D 2016
+%T Characteristics of patients with type 2 diabetes mellitus newly treated with GLP-1 receptor agonists (CHADIG Study): A cross-sectional multicentre study in Spain
+%B BMJ Open. 6 (7) (no pagination), 2016. Article Number: e010197. Date of Publication: 01 Jul 2016.
+%! Characteristics of patients with type 2 diabetes mellitus newly treated with GLP-1 receptor agonists (CHADIG Study): A cross-sectional multicentre study in Spain
+%R http://dx.doi.org/10.1136/bmjopen-2015-010197
+%M CN-01196504
+%X Objective: Several glucagon-like peptide-1 (GLP-1) receptor agonists (GLP-1Ra) have been made recently available in Spain for type 2 diabetes mellitus (DM2) treatment. There are no published data on the clinical and sociodemographic profile of patients initiating treatment with GLP-1Ra in Spain. Our objective was to understand these patients' characteristics in a real-world clinical practice setting. Design: Cross-sectional observational study. Setting: Spanish specialist outpatient clinics. Participants: 403 adults with DM2 initiating GLP-1Ra treatment were included. Primary and secondary outcome measures: Sociodemographic and DM2-related clinical data, including treatment at and after GLP-1Ra initiation and comorbidities, were collected. Results: Evaluable patients (n=403; 50.9% female) were included (July 2013 to March 2014) at 24 centres by 53 specialists (47 endocrinology, 6 internal medicine), with the following profile (value+/-SD): age (58.3+/-10.4 years), diabetes duration (9.9+/-7 years), body mass index (BMI; 36.2+/-5.5) and glycated haemoglobin (HbA1c; 8.4+/-1.4%); 14% had HbA1c<7%. Previous antidiabetic treatment: 53.8% only oral antidiabetic drugs (OADs), 5.2% insulin and 40% insulin and OAD; of those receiving OAD, 35% single drug, 38.2% 2 drugs and 24% 3 drugs. Concomitant to GLP-1Ra, 55.3% were only on OAD, 36.2% on insulin and OAD, and 7.2% only on insulin. Of those receiving OAD, the GLP-1Ra was mainly associated with 1 drug (65%) or 2 drugs (31.8%). GLP-1Ra are frequently added to existing antidiabetic drugs, with dipeptidyl peptidase-4 inhibitors being the OAD most frequently switched (45% receiving 1 before starting GLP-1Ra, only 2.7% receiving it concomitantly). Conclusions: In Spain, GLP-1Ra therapy is usually started in combination with OADs or OADs and insulin. These drugs are used in relatively young patients often not reaching therapeutic goals with other treatment combinations, roughly a decade after diagnosis and with a relatively high BMI. The latter could be explained by Spanish regional payers limiting reimbursed prescription to patients with a minimum BMI threshold (>30 in most regions, >35 in some). Copyright © 2016, BMJ Publishing Group. All rights reserved.
+%~ CENTRAL
+%G eng
+
+
+
+%0 Journal Article
+%D 2016
+%T Effects of vildagliptin relative to sulfonylureas in muslim patients with type 2 diabetes fasting during ramadan: Influence of age and treatment with/without metformin in the VIRTUE study
+%B Diabetes, Metabolic Syndrome and Obesity: Targets and Therapy. 9 (pp 225-231), 2016. Date of Publication: 26 Jul 2016.
+%! Effects of vildagliptin relative to sulfonylureas in muslim patients with type 2 diabetes fasting during ramadan: Influence of age and treatment with/without metformin in the VIRTUE study
+%R http://dx.doi.org/10.2147/DMSO.S103692
+%M CN-01194158
+%X Background: VIRTUE was a prospective, observational study assessing the effectiveness and safety of vildagliptin vs sulfonylureas (SUs) (both as monotherapy and in combination with metformin) in patients with type 2 diabetes mellitus who fasted during Ramadan. A post hoc analysis was carried out to assess the effect of treatment with/without metformin and age (<65 years or >65 years). Patients and methods: Patients were recruited from the Middle East and Asia. The primary end point was proportion of patients with one or more hypoglycemic event (HE) during Ramadan. Secondary end points included change from baseline in glycated hemoglobin (HbA1c), body weight, and safety. Results: Overall, 684 patients received vildagliptin and 631 received SUs. Most patients received dual therapy with metformin (n=1,148) and were aged,65 years (n=1,189). A few patients experienced one or more HE with vildagliptin vs SU monotherapy (6.5% vs 14.5%) and with vildagliptin +/- metformin vs SUs + metformin (5.3% vs 20.6%); the latter achieved statistical significance (P,0.001) in both age subgroups (<65 years: 5.5% vs 18.4%, P,0.001; >65 years: 2.8% vs 30.9%, P,0.001). Vildagliptin was associated with numerically greater HbA1c and body weight reductions vs SUs, regardless of the therapy type or age. A higher proportion of SU-vs vildagliptin-treated patients experienced adverse events across all subgroups. Conclusion: A few patients experienced HEs with vildagliptin vs SUs regardless of age, and in patients on dual therapy. Vildagliptin +/- metformin was also associated with good glycemic and weight control and was well tolerated. Vildagliptin might be a useful treatment option for patients with type 2 diabetes mellitus, particularly high-risk populations such as the elderly fasting during Ramadan. Copyright © 2016 Hassoun et al.
+%~ CENTRAL
+%G eng
+
+
+
+%0 Journal Article
+%D 2016
+%T The relationship of management modality in Saudi patients with type 2 diabetes to components of metabolic syndrome, gamma glutamyl transferase and highly sensitive C-reactive protein
+%B Therapeutic Advances in Chronic Disease. 7 (5) (pp 246-254), 2016. Date of Publication: 01 Sep 2016.
+%! The relationship of management modality in Saudi patients with type 2 diabetes to components of metabolic syndrome, gamma glutamyl transferase and highly sensitive C-reactive protein
+%R http://dx.doi.org/10.1177/2040622316658459
+%M CN-01194266
+%X Background: The aim of this study was to investigate the relationship among management modality, glycemic control, components of metabolic syndrome (MS) and serum levels of gamma glutamyl transferase (GGT) and C-reactive protein (CRP) in patients with type 2 diabetes (T2DM). Methods: Patients with T2DM, not suffering from diabetes complications, were recruited from outpatients clinics at two hospitals in Jeddah. Anthropometric measurements and blood pressure (BP) were taken. A treatment plan was recorded. Fasting blood samples were obtained to measure glucose, glycated hemoglobin (HbA1c), lipids profile, highly sensitive (hs)-CRP and GGT. Results: A total of 71 men and 82 women were recruited. Lower mean HbA1c was found in people receiving oral glucose-lowering drugs compared with those on insulin therapy (p < 0.001). Management modality had no effect on mean GGT or hs-CRP. Higher mean GGT was associated with poor glycemic control, dyslipidemia, hypertension, and abdominal obesity. GGT correlated significantly (p < 0.05) and directly with triglycerides in men (r = 0.401) and diastolic BP (r = 0.279 for men, r = 0.194, for women), but inversely with high-density lipoprotein cholesterol (HDL-C) (r = -0.298 for men, r = -0.171 for women). hs-CRP correlated with waist circumference (p < 0.05, r = 0.312, for men, r = 0.305, for women), with a higher mean being found in men with poor glycemic control (p = 0.015), in hypertensive women (p = 0.030), and in patients who were abdominally obese (p < 0.05). Conclusions: High levels of GGT and hs-CRP are associated with components of MS and poor glycemic control, hence increased cardiovascular risk. Due to their value as independent risk predictors of vascular injury, these measures should be included in routine monitoring of patients with T2DM. Copyright © 2016, © The Author(s), 2016.
+%~ CENTRAL
+%G eng
+`
+
+export default dummypapers;
