+++ src/PaperScreening/Components/SideBar/Containers/LibraryCounts.js
@@ -1,10 +1,10 @@
 import { Checkbox, Panel } from 'react-bootstrap';
-import { Colors, Decision } from '../../../../../Elements/constants';
+import { Colors, Decision } from '../../../../Elements/constants';
 import React, { Component } from 'react';
 
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { updateFilter } from '../../../../../Actions';
+import { updateFilter } from '../../../../Actions';
 
 class LibraryCounts extends Component {
     constructor(props, context) {
